Add support for ZOBEX DD-FDC disks (8", DSDD/SSDD, CP/M 2.2)
This work was based on flux images shared by Snekers on the Tech Tangents
Discord, which contained CP/M 2.2 and a fair chunk of documentation and
source code relating to ZOBEX DD-FDC disk controller.  This included the
CP/M 2.2 BIOS, from which the cpmtools diskdefs referenced in the two
decode scripts were derived; the DSDD one has been tested against three
flux images successfully.  The SSDD one is as-yet untested.