Embed custom CP/M diskdefs into scripts that need them.
This avoids requiring the user to add the various diskdefs to their system-wide
/usr/local/share/diskdefs file, by passing the definitions directly to cpmtools.

I've tried to be consistent in layout here - the diskdef is always in a variable
called cpm_diskdef, includes a name and is preceded by a comment identifying the
system it belongs to.  Before that is a blank line.  There's method to this
madness...